Correções
Remoção da entidade tempo, usuarios só recuperam seus proprios dados.

+++ src/main/java/com/allos/pomodoro/service/TarefasService.java
@@ -7,7 +7,6 @@ import com.allos.pomodoro.exception.ObjectNotFoundException;
 import com.allos.pomodoro.mapper.TarefasMapper;
 import com.allos.pomodoro.repository.TarefasRepository;
 import com.allos.pomodoro.security.UserSecurity;
-import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -42,16 +41,32 @@ public class TarefasService {
     }
 
     public TarefasDTO atualizarTarefa(TarefasDTO dto){
+
+        UserSecurity userSecurity = UserSecurityService.authenticate();
+
+        if(userSecurity == null || !userSecurity.hasRole(Perfil.ADMIN) && dto.getId().equals(userSecurity.getId())){
+            throw new AuthorizationException("Acesso negado");
+        }
+
         if(!repository.findById(dto.getId()).isPresent()){
             throw new ObjectNotFoundException("Tarefa não existe.");
         }
+
         return mapper.toDto(repository.save(mapper.toEntity(dto)));
     }
 
     public void deletaTarefa(Long id) {
+
+        UserSecurity userSecurity = UserSecurityService.authenticate();
+
+        if(userSecurity == null || !userSecurity.hasRole(Perfil.ADMIN) && !id.equals(userSecurity.getId())){
+            throw new AuthorizationException("Acesso negado");
+        }
+
         if(!repository.findById(id).isPresent()){
             throw new ObjectNotFoundException("Usuario não existe.");
         }
         repository.deleteById(id);
     }
+
 }
